Fix a few things (compiler issue on previous PETSc, parallel test of !GetVecValue and type mis-match).
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@3951 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ heart/src/problem/MonodomainFastSlowProblem.hpp
@@ -91,7 +91,7 @@ public:
      */
     MonodomainFastSlowProblem(AbstractCardiacCellFactory<DIM>* pCellFactory,
 				      		  MixedTetrahedralMesh<DIM,DIM>& rMixedMesh,
-						      unsigned slowCellsTimeStep,
+						      double slowCellsTimeStep,
         		              bool orthotropicMedia=true)
             : MonodomainProblem<DIM>(pCellFactory, orthotropicMedia),
               mrMixedMesh(rMixedMesh),
